requirements.txt
gradio
tqdm
open_clip_torch
torch
pillow
numpy
opencv-python
imagehash